Updated CTexture to store its members as pointers, NOT as raw values.
If we don't do this, then the vectory can get reallocated and the pointers from all the entities that use it will become invalid and require a search again. To solve this we weither had to store an iterator (which I don't know would work) or use pointers (as their value is all we care about NOT what's in them!